fm-225: Code clean up by consolidating all reusable functions into util.ts and all constant variables to constant.ts
