+++ src/components/play-button.ts
@@ -1,5 +1,5 @@
 import { TappedStart } from "../Firebase/firebase-event-interface";
-import { loadImages } from "../common/common";
+import { loadImages } from "../common/";
 import { FirebaseIntegration } from "../Firebase/firebase-integration";
 import { lang, pseudoId } from "../../global-variables";
 import { getData } from "../data/api-data";
